Update smb3.asl
Adds a new method for end-of-level splits and removes the broken one.

This has been tested and appears working by watching the fanfare queue byte at `0x4e4`, so it will correctly split normal stage and fortress exits by watching for the appropriate music to begin playing. AFAICT, this is frame-perfect.